pyproject.toml
[project]
name = "catocode"
version = "0.1.0"
description = "AI-powered GitHub issue reproducer, fixer, and verifier with evidence chains"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pyjwt>=2.11.0",
    "cryptography>=46.0.5",
]

[project.scripts]
catocode = "catocode.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/catocode"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
